requirements.txt
annotated-types==0.7.0
anyio==4.4.0
Authlib==1.3.2
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
contourpy==1.3.0
cryptography==43.0.1
cycler==0.12.1
exceptiongroup==1.2.2
fastapi==0.114.2
filelock==3.16.0
fonttools==4.53.1
fsspec==2024.9.0
grpcio==1.66.1
grpcio-health-checking==1.66.1
grpcio-tools==1.66.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.24.7
idna==3.9
Jinja2==3.1.4
joblib==1.4.2
kiwisolver==1.4.7
MarkupSafe==2.1.5
matplotlib==3.9.2
mpmath==1.3.0
networkx==3.3
numpy==1.26.4
packaging==24.1
pillow==10.4.0
protobuf==5.28.1
pycparser==2.22
pydantic==2.9.1
pydantic_core==2.23.3
pyparsing==3.1.4
python-dateutil==2.9.0.post0
PyYAML==6.0.2
regex==2024.9.11
requests==2.32.3
safetensors==0.4.5
scikit-learn==1.5.2
scipy==1.14.1
sentence-transformers==3.1.0
six==1.16.0
sniffio==1.3.1
starlette==0.38.5
sympy==1.13.2
threadpoolctl==3.5.0
tokenizers==0.19.1
torch==2.2.2
tqdm==4.66.5
transformers==4.44.2
typing_extensions==4.12.2
urllib3==2.2.3
uvicorn==0.30.6
validators==0.28.1
weaviate-client==4.6.0
